code refractor #1
