+++ src/main/java/com/flw/moka/entity/helpers/ProviderResponseData.java
@@ -1,4 +1,4 @@
-package com.flw.moka.entity;
+package com.flw.moka.entity.helpers;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -11,14 +11,13 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProviderResponse {
-
-    @JsonProperty("Data")
-    private ProviderResponseData Data;
+public class ProviderResponseData {
+    @JsonProperty("IsSuccessful")
+    private String IsSuccessful;
     @JsonProperty("ResultCode")
     private String ResultCode;
     @JsonProperty("ResultMessage")
     private String ResultMessage;
-    @JsonProperty("Exception")
-    private String Exception;
+    @JsonProperty("VirtualPosOrderId")
+    private String VirtualPosOrderId;
 }
