+++ src/main/java/com/flw/moka/entity/helpers/ProxyResponse.java
@@ -1,4 +1,4 @@
-package com.flw.moka.entity;
+package com.flw.moka.entity.helpers;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
