+++ src/main/java/com/flw/moka/repository/ProviderPayloadRepository.java
@@ -2,9 +2,9 @@ package com.flw.moka.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.flw.moka.entity.PaymentDealerAuthentication;
-import com.flw.moka.entity.PaymentDealerRequest;
-import com.flw.moka.entity.ProviderPayload;
+import com.flw.moka.entity.helpers.PaymentDealerAuthentication;
+import com.flw.moka.entity.helpers.PaymentDealerRequest;
+import com.flw.moka.entity.helpers.ProviderPayload;
 
 @Repository
 public class ProviderPayloadRepository {
