+++ src/main/java/com/flw/moka/service/controllers/CaptureService.java
@@ -10,12 +10,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.flw.moka.entity.CardParams;
-import com.flw.moka.entity.ProductRequest;
-import com.flw.moka.entity.ProviderPayload;
-import com.flw.moka.entity.ProviderResponse;
-import com.flw.moka.entity.ProviderResponseData;
-import com.flw.moka.entity.ProxyResponse;
 import com.flw.moka.entity.Transaction;
+import com.flw.moka.entity.helpers.ProductRequest;
+import com.flw.moka.entity.helpers.ProviderPayload;
+import com.flw.moka.entity.helpers.ProviderResponse;
+import com.flw.moka.entity.helpers.ProviderResponseData;
+import com.flw.moka.entity.helpers.ProxyResponse;
 import com.flw.moka.service.entities.CardParamsService;
 import com.flw.moka.service.entities.ProxyResponseService;
 import com.flw.moka.service.entities.TransactionService;
@@ -79,9 +79,6 @@ public class CaptureService {
         transaction.setTransactionRef(productRequest.getTransactionReference());
         transaction.setTransactionStatus("capture");
         transaction.setTimeCaptured(timeUtility.getDateTime());
-
-        // transaction.setTimeCaptured("2023-03-16 12:55:22");
-        // test void-refund logic with "2023-03-16 12:55:22"
 
         return transaction;
     }
