+++ src/main/java/com/flw/moka/service/entities/CardParamsServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.flw.moka.entity.CardParams;
-import com.flw.moka.entity.ProxyResponse;
+import com.flw.moka.entity.helpers.ProxyResponse;
 import com.flw.moka.exception.TransactionAlreadyCapturedException;
 import com.flw.moka.exception.TransactionNotFoundException;
 import com.flw.moka.repository.CardParamsRepository;
