+++ src/main/java/com/flw/moka/service/entities/PaymentDealerRequestService.java
@@ -1,7 +1,7 @@
 package com.flw.moka.service.entities;
 
-import com.flw.moka.entity.PaymentDealerRequest;
-import com.flw.moka.entity.ProductRequest;
+import com.flw.moka.entity.helpers.PaymentDealerRequest;
+import com.flw.moka.entity.helpers.ProductRequest;
 
 public interface PaymentDealerRequestService {
     PaymentDealerRequest saveRequestPayload(ProductRequest productRequest, String method);
