+++ src/main/java/com/flw/moka/service/entities/ProviderPayloadService.java
@@ -1,7 +1,7 @@
 package com.flw.moka.service.entities;
 
-import com.flw.moka.entity.PaymentDealerRequest;
-import com.flw.moka.entity.ProviderPayload;
+import com.flw.moka.entity.helpers.PaymentDealerRequest;
+import com.flw.moka.entity.helpers.ProviderPayload;
 
 public interface ProviderPayloadService {
     ProviderPayload savePaymentDealerAuthAndReq(PaymentDealerRequest paymentDealerRequest);
