+++ src/main/java/com/flw/moka/service/entities/ProviderPayloadServiceImpl.java
@@ -2,9 +2,9 @@ package com.flw.moka.service.entities;
 
 import org.springframework.stereotype.Service;
 
-import com.flw.moka.entity.PaymentDealerAuthentication;
-import com.flw.moka.entity.PaymentDealerRequest;
-import com.flw.moka.entity.ProviderPayload;
+import com.flw.moka.entity.helpers.PaymentDealerAuthentication;
+import com.flw.moka.entity.helpers.PaymentDealerRequest;
+import com.flw.moka.entity.helpers.ProviderPayload;
 import com.flw.moka.repository.PaymentDealerAuthenticationRepository;
 import com.flw.moka.repository.ProviderPayloadRepository;
