+++ src/main/java/com/flw/moka/service/entities/ProxyResponseService.java
@@ -2,10 +2,10 @@ package com.flw.moka.service.entities;
 
 import java.util.Optional;
 
-import com.flw.moka.entity.ProductRequest;
-import com.flw.moka.entity.ProviderResponse;
-import com.flw.moka.entity.ProviderResponseData;
-import com.flw.moka.entity.ProxyResponse;
+import com.flw.moka.entity.helpers.ProductRequest;
+import com.flw.moka.entity.helpers.ProviderResponse;
+import com.flw.moka.entity.helpers.ProviderResponseData;
+import com.flw.moka.entity.helpers.ProxyResponse;
 
 public interface ProxyResponseService {
     ProxyResponse createProxyResponse(Optional<ProviderResponseData> dataEntity, Optional<ProviderResponse> bodyEntity,
